fix: reflect feedback
필터 파일 분리
에러 수정
